requirements.txt
mediapipe
opencv-python
pytorch_lightning